+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-alpha8] - 2019-02-18
+ ### Add
+ - `ResponseError` class now contains the occurred exception.
+ - `BadGatewayHttpException`
+ ### Fix
+ - URL encoding fix.
+ ### Change
+ - **[BC BREAK]** `HttpException` `newInstance` function will return with the exception object, not throwing it.
+
 ## [1.0.0-alpha7] - 2019-02-14
  ### Add
  - `shardimage\shardimagephpapi\services\Client` has `timeout` property from now.